refactor: tests layout
